Create LICENSE
Added a LICENSE file to the repository using the MIT License. This inclusion ensures that the project is open-source and freely available for public use, modification, and distribution under defined conditions. The license clearly outlines user rights and limitations, providing legal protection and transparency for the project and its contributors